+++ src-tauri/yaak-plugins/src/nodejs.rs
@@ -9,12 +9,6 @@ use tauri_plugin_shell::ShellExt;
 use tauri_plugin_shell::process::CommandEvent;
 use tokio::sync::watch::Receiver;
 
-#[derive(Deserialize, Default)]
-#[serde(default, rename_all = "camelCase")]
-struct PortFile {
-    port: i32,
-}
-
 pub async fn start_nodejs_plugin_runtime<R: Runtime>(
     app: &AppHandle<R>,
     addr: SocketAddr,
