+++ src-web/components/FolderSettingsDialog.tsx
@@ -1,13 +1,17 @@
-import { foldersAtom, patchModel } from '@yaakapp-internal/models';
+import { createWorkspaceModel, foldersAtom, patchModel } from '@yaakapp-internal/models';
 import { useAtomValue } from 'jotai';
 import { useMemo, useState } from 'react';
 import { useAuthTab } from '../hooks/useAuthTab';
+import { useEnvironmentsBreakdown } from '../hooks/useEnvironmentsBreakdown';
 import { useHeadersTab } from '../hooks/useHeadersTab';
 import { useInheritedHeaders } from '../hooks/useInheritedHeaders';
+import { Button } from './core/Button';
 import { Input } from './core/Input';
 import { VStack } from './core/Stacks';
 import type { TabItem } from './core/Tabs/Tabs';
 import { TabContent, Tabs } from './core/Tabs/Tabs';
+import { EmptyStateText } from './EmptyStateText';
+import { EnvironmentEditor } from './EnvironmentEditor';
 import { HeadersEditor } from './HeadersEditor';
 import { HttpAuthenticationEditor } from './HttpAuthenticationEditor';
 import { MarkdownEditor } from './MarkdownEditor';
@@ -19,6 +23,7 @@ interface Props {
 
 const TAB_AUTH = 'auth';
 const TAB_HEADERS = 'headers';
+const TAB_VARIABLES = 'variables';
 const TAB_GENERAL = 'general';
 
 export type FolderSettingsTab = typeof TAB_AUTH | typeof TAB_HEADERS | typeof TAB_GENERAL;
@@ -30,6 +35,17 @@ export function FolderSettingsDialog({ folderId, tab }: Props) {
   const authTab = useAuthTab(TAB_AUTH, folder);
   const headersTab = useHeadersTab(TAB_HEADERS, folder);
   const inheritedHeaders = useInheritedHeaders(folder);
+  const environments = useEnvironmentsBreakdown();
+  const folderEnvironment = environments.allEnvironments.find(
+    (e) => e.parentModel === 'folder' && e.parentId === folderId,
+  );
+
+  const handleActiveTabChange = (tab: string) => {
+    setActiveTab(tab);
+    if (tab === TAB_VARIABLES && folderEnvironment == null) {
+      console.log('CREATE FOLDER ENVIRONMENT');
+    }
+  };
 
   const tabs = useMemo<TabItem[]>(() => {
     if (folder == null) return [];
@@ -39,8 +55,12 @@ export function FolderSettingsDialog({ folderId, tab }: Props) {
         value: TAB_GENERAL,
         label: 'General',
       },
-      ...authTab,
       ...headersTab,
+      ...authTab,
+      {
+        value: TAB_VARIABLES,
+        label: 'Variables',
+      },
     ];
   }, [authTab, folder, headersTab]);
 
@@ -49,7 +69,7 @@ export function FolderSettingsDialog({ folderId, tab }: Props) {
   return (
     <Tabs
       value={activeTab}
-      onChangeValue={setActiveTab}
+      onChangeValue={handleActiveTabChange}
       label="Folder Settings"
       className="px-1.5 pb-2"
       addBorders
@@ -84,6 +104,31 @@ export function FolderSettingsDialog({ folderId, tab }: Props) {
           onChange={(headers) => patchModel(folder, { headers })}
           stateKey={`headers.${folder.id}`}
         />
+      </TabContent>
+      <TabContent value={TAB_VARIABLES} className="pt-3 overflow-y-auto h-full px-4">
+        {folderEnvironment == null ? (
+          <EmptyStateText>
+            <VStack alignItems="center" space={1.5}>
+              <p>Folder environment not found for folder</p>
+              <Button
+                variant="border"
+                size="sm"
+                onClick={async () => {
+                  await createWorkspaceModel({
+                    workspaceId: folder.workspaceId,
+                    parentModel: 'folder',
+                    parentId: folder.id,
+                    model: 'environment',
+                  });
+                }}
+              >
+                Create Folder Environment
+              </Button>
+            </VStack>
+          </EmptyStateText>
+        ) : (
+          <EnvironmentEditor environment={folderEnvironment} />
+        )}
       </TabContent>
     </Tabs>
   );
