+++ src-web/hooks/useEnvironmentsBreakdown.ts
@@ -6,7 +6,8 @@ export function useEnvironmentsBreakdown() {
   const allEnvironments = useAtomValue(environmentsAtom);
   return useMemo(() => {
     const baseEnvironments = allEnvironments.filter((e) => e.parentId == null) ?? [];
-    const subEnvironments = allEnvironments.filter((e) => e.parentId != null) ?? [];
+    const subEnvironments =
+      allEnvironments.filter((e) => e.parentModel === 'environment' && e.parentId != null) ?? [];
 
     const baseEnvironment = baseEnvironments[0] ?? null;
     const otherBaseEnvironments =
